+++ src/client/public/index.ts
@@ -1,4 +1,4 @@
 export { SuperfaceClient, createTypedClient, typeHelper } from './client';
-export { Profile, ProfileConfiguration } from './profile';
+export { Profile, ProfileConfiguration, TypedProfile } from './profile';
 export { Provider, ProviderConfiguration } from './provider';
 export { UseCase } from './usecase';